feat: 지하철 노선 구간 삭제 시에 마지막 하행역이 아닐 시 제거 할 수 없다

+++ src/main/java/subway/application/service/SectionCommandService.java
@@ -6,10 +6,7 @@ import subway.application.service.input.LineLoadUseCase;
 import subway.application.service.input.SectionCommandUseCase;
 import subway.application.service.output.SectionCommandRepository;
 import subway.application.service.output.SectionLoadRepository;
-import subway.domain.Line;
-import subway.domain.Section;
-import subway.domain.SectionCreateDto;
-import subway.domain.Station;
+import subway.domain.*;
 
 import java.util.List;
 
@@ -45,8 +42,14 @@ public class SectionCommandService implements SectionCommandUseCase {
     @Override
     public void deleteLineSection(Long lineId, Long sectionId) {
         Section section = sectionLoadRepository.loadLineSectionWithLineId(lineId, sectionId);
-        sectionCommandRepository.deleteSection(section.getId());
+        List<Section> sections = sectionLoadRepository.loadLineSection(lineId);
+        Section lastSection = sections.get(sections.size() - 1);
 
+        if (lastSection.getDownStation().equals(section.getUpStation())) {
+            sectionCommandRepository.deleteSection(section.getId());
+        }
+
+        throw new NotDownStationWhenDeleteSectionException();
     }
 
 }
